Moving and separating dataclasses from Dao interfaces
