+++ app/src/main/java/pl/llp/aircasting/data/api/repository/MeasurementsRepository.kt
@@ -1,7 +1,7 @@
 package pl.llp.aircasting.data.api.repository
 
 import pl.llp.aircasting.data.local.DatabaseProvider
-import pl.llp.aircasting.data.local.data_classes.MeasurementDBObject
+import pl.llp.aircasting.data.local.entity.MeasurementDBObject
 import pl.llp.aircasting.data.model.Measurement
 import java.util.*
