+++ app/src/main/java/pl/llp/aircasting/data/local/entity/ExtSession.kt
@@ -1,0 +1,13 @@
+package pl.llp.aircasting.data.local.entity
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class ExtSessionEntity(
+    @PrimaryKey val uid: Int,
+    @ColumnInfo(name = "first_name") val firstName: String?,
+    @ColumnInfo(name = "last_name") val lastName: String?
+)
+
