+++ app/src/main/java/pl/llp/aircasting/data/model/Measurement.kt
@@ -1,7 +1,7 @@
 package pl.llp.aircasting.data.model
 
-import pl.llp.aircasting.data.local.data_classes.ActiveSessionMeasurementDBObject
-import pl.llp.aircasting.data.local.data_classes.MeasurementDBObject
+import pl.llp.aircasting.data.local.entity.ActiveSessionMeasurementDBObject
+import pl.llp.aircasting.data.local.entity.MeasurementDBObject
 import pl.llp.aircasting.util.events.NewMeasurementEvent
 import pl.llp.aircasting.util.DateConverter
 import pl.llp.aircasting.data.api.response.MeasurementResponse
