+++ app/src/main/java/pl/llp/aircasting/ui/viewmodel/SessionsViewModel.kt
@@ -2,7 +2,7 @@ package pl.llp.aircasting.ui.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import pl.llp.aircasting.data.local.data_classes.*
+import pl.llp.aircasting.data.local.entity.*
 import pl.llp.aircasting.data.local.DatabaseProvider
 import pl.llp.aircasting.data.model.MeasurementStream
 import pl.llp.aircasting.data.model.SensorThreshold
