+++ app/src/main/java/pl/llp/aircasting/util/helpers/sensor/airbeam3/sync/SDCardFixedSessionsProcessor.kt
@@ -1,7 +1,7 @@
 package pl.llp.aircasting.util.helpers.sensor.airbeam3.sync
 
 import pl.llp.aircasting.data.local.DatabaseProvider
-import pl.llp.aircasting.data.local.data_classes.MeasurementDBObject
+import pl.llp.aircasting.data.local.entity.MeasurementDBObject
 import pl.llp.aircasting.data.api.repository.MeasurementStreamsRepository
 import pl.llp.aircasting.data.api.repository.MeasurementsRepository
 import pl.llp.aircasting.data.api.repository.SessionsRepository
